refactor(ui): extract the tool call content view
